Move class to right package
