+++ mobile/src/test/java/com/alexstyl/specialdates/events/namedays/calendar/RomanianEasterSpecialCalculatorTest.java
@@ -1,7 +1,6 @@
-package com.alexstyl.specialdates.events.namedays;
+package com.alexstyl.specialdates.events.namedays.calendar;
 
 import com.alexstyl.specialdates.date.DayDate;
-import com.alexstyl.specialdates.events.namedays.calendar.RomanianEasterSpecialCalculator;
 
 import java.util.ArrayList;
 import java.util.List;
